+++ R/utils.R
@@ -49,7 +49,7 @@ city_to_url <- function(city_, feed_) {
       return(city_)
     } else {
       stop(sprintf(c("The supplied argument for \"city\" looks like a URL, ",
-                     "but the webpage doesn't seem to exist. Please check",
+                     "but the webpage doesn't seem to exist. Please check ",
                      "the URL provided or provide the city name as a string.")))
     }
   }
@@ -161,6 +161,11 @@ determine_output_types <- function(directory_, output_) {
 
 # a function that grabs a gbfs formatted dataset
 get_gbfs_dataset_ <- function(city, directory, file, output, feed) {
+  
+  # test internet connection
+  if (!connected_to_internet()) {
+    return(message_no_internet())
+  }
   
   # check arguments to make sure the putput method makes sense
   check_return_arguments(directory_ = directory,
@@ -342,10 +347,13 @@ url_exists <- function(x, quiet = FALSE, ...) {
   sHEAD <- safely(httr::HEAD)
   sGET <- safely(httr::GET)
   
+  if (!stringr::str_detect(x, "http")) {
+    x <- paste0("https://", x)
+  }
+  
   res <- sHEAD(x, ...)
   
-  if (is.null(res$result) || 
-      ((httr::status_code(res$result) %/% 200) != 1)) {
+  if (is.null(res$result)) {
     
     res <- sGET(x, ...)
     
@@ -356,4 +364,17 @@ url_exists <- function(x, quiet = FALSE, ...) {
   
   return(TRUE)
   
+}
+
+# a function to alert the user of no internet connection in a
+# more informative/helpful way
+message_no_internet <- function() {
+  message(c("You don't seem to have an active internet connection. Please", 
+            "connect to the internet to use the gbfs package."))
+  return(list())
+}
+
+# a wrapper around has internet so that with_mock can be used in tests
+connected_to_internet <- function() {
+  curl::has_internet()
 }